settings(F1.2): relocate 4 legacy profile modules into sections/llm/
ai-profiles/tts-profiles/stt-profiles/vision-profiles were unregistered
sub-modules of the llm section but sat in sections/ looking like standalone
menu sections. Move them under sections/llm/ (their sole importer,
profiles-tab.js) so ownership is unambiguous; profiles-tab now imports
them via ./. Relative specifiers (incl. the vision test-fixture repo-root
path) gain one ../ for the deeper location. Inlining ~1650 lines into
profiles-tab.js was rejected as a megafile; relocation achieves the same
single-source-of-truth goal. No behavior change.
